+++ src/client/java/tempeststudios/inventorysort/mixin/AbstractRecipeBookScreenMixin.java
@@ -1,5 +1,6 @@
 package tempeststudios.inventorysort.mixin;
 
+import net.minecraft.client.gui.GuiGraphics;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.injection.At;
@@ -11,9 +12,10 @@ import tempeststudios.inventorysort.RecipeBookAwareButtonScreen;
 @Mixin(targets = "net.minecraft.client.gui.screens.inventory.AbstractRecipeBookScreen")
 public abstract class AbstractRecipeBookScreenMixin {
 	@Inject(method = "render", at = @At("HEAD"), require = 0)
-	private void inventorySort$updateButtonsOnRecipeBookRender(CallbackInfo ci) {
+	private void inventorySort$updateButtonsOnRecipeBookRender(GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick, CallbackInfo ci) {
 		if (this instanceof RecipeBookAwareButtonScreen buttonScreen) {
 			buttonScreen.inventorysort$updateButtonPositionsFromRecipeBookRender();
+			buttonScreen.inventorysearch$updateButtonPositionsFromRecipeBookRender();
 		}
 	}
 }